// medflow-erp/types.ts
export interface Bed {
  id: string;
  number: string;
  status: 'Occupied' | 'Cleaning' | 'Available';
  patientName?: string;
  daysSinceAdmission?: number;
}